reorganized .h files
