+++ example/example4.cpp
@@ -1,5 +1,5 @@
-#include "../src/RecordLoader.h"
-#include "../src/QueryProcessor.h"
+#include "../src/lib/RecordLoader.h"
+#include "../src/lib/QueryProcessor.h"
 
 int main() {
     char* file_path = "../dataset/bestbuy_sample_small_records.json";
