+++ src/RecordLoader.cpp
@@ -1,5 +1,5 @@
 #include <sys/time.h>
-#include "RecordLoader.h"
+#include "./lib/RecordLoader.h"
 using namespace std;
 
 #define MAX_PAD 64
