+++ src/lib/QueryAutomaton.h
