+++ src/lib/QueryProcessor.h
